refactor(models): move estimator to models/utils; add shim at ml_playground/estimator.py
